+++ CHANGES.md
@@ -9,6 +9,7 @@
         to get a proof clean, you need to resolve all the proofs returned by this call.
       - Add a knuckle-dragger proof that tail-recursive definition of fibonacci is
         equivalent to its textbook definition. See Documentation/SBV/Examples/KnuckleDragger/Fibonacci.hs.
+      - Renamed kdShowDepsHTML to showProofTreeHTML. (Along with showProofTree which renders in ASCII.)
 
   * Remove Data.SBV.Tools.NaturalInduction. The functionality provided by this tool is much better
     addressed by KnuckleDragger's induction methods. If you were using this functionality and have problems